requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
flask-cors
flask-limiter
python-dotenv
requests
beautifulsoup4
google-generativeai
groq
sumy
